Generalized generator to create TypeScript declarations as well.
