+++ lib/generate.js
@@ -204,6 +204,7 @@ var main = module.exports = function() {
     });
 
     require('./generateFlowTypes.js');
+    require('./generateTypeScriptTypes.js');
 };
 
 main();
