+++ lib/generateTypeScriptTypes.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+module.exports = require("./generateTypes")("TypeScript", "index.d.ts.tpl", "index.d.ts");
